+++ src/react/views/App.js
@@ -36,7 +36,7 @@ export default class App extends React.Component {
     if (!this.state.stateHandler) return null
     const locale = this.state.stateHandler.locale
     return <IntlProvider locale={locale} messages={locales[locale].messages} textComponent={Text}>
-        <View style={app.mainContainer}>{this.renderContent()}</View>
+      <View style={app.mainContainer}>{this.renderContent()}</View>
     </IntlProvider>
   }
 }