+++ src/react/views/AuthWebView.js
@@ -1,9 +1,8 @@
 // @flow
 import React from 'react'
 import LoadingWebView from './LoadingWebView'
-import type { WebView } from 'react-native'
-import config from '../../config'
 import { authWebView } from '../../style'
+import { Linking } from 'react-native'
 import url from 'url'
 
 class AuthWebView extends React.Component {
@@ -15,81 +14,51 @@ class AuthWebView extends React.Component {
     onSuccess: () => void,
     onAuthFailed: () => void
   }
-  interval: number
-  _messageHandled: boolean
-  ref: WebView
+  handle = (evt: { url: string }) => this._handleOpenURL(evt)
 
-  isBack (u: string) {
-    const backUrl = `${config.laundree.host}/native-app`
-    const {host, auth, protocol, path} = url.parse(backUrl)
-    const {host: host2, auth: auth2, protocol: protocol2, path: path2} = url.parse(u)
-    return auth === auth2 && protocol === protocol2 && path === path2 && host === host2
-  }
-
-  onLoadUrl (url: string) {
-    if (!this.isBack(url)) return
-    this.startSendTokenRequest()
-  }
-
-  injectedJavaScript = `(function() {
-  var originalPostMessage = window.postMessage;
-  var patchedPostMessage = function(message, targetOrigin, transfer) { 
-    originalPostMessage(message, targetOrigin, transfer);
-  };
-  patchedPostMessage.toString = function() { 
-    return String(Object.hasOwnProperty).replace('hasOwnProperty', 'postMessage'); 
-  };  
-  window.postMessage = patchedPostMessage;
-})();`
-
-  handleMessage (message: string) {
-    console.log('Handling message', message)
-    if (this._messageHandled) return
-    console.log('Handling message', message)
-    this._messageHandled = true
-    this.stopSendTokenRequest()
-    try {
-      const {secret, userId} = JSON.parse(message)
-      if (!(secret && userId)) {
-        console.log('No token provided. Auth failed.')
-        return this.props.onAuthFailed()
-      }
-      this.props.onSuccess({secret, userId})
-    } catch (error) {
-      console.log(error)
-    }
-  }
-
-  startSendTokenRequest () {
-    console.log('Starting send token request')
-    this.sendTokenRequest()
-    this.interval = setInterval(() => this.sendTokenRequest(), 100)
-  }
-
-  stopSendTokenRequest () {
-    if (!this.interval) return
-    clearInterval(this.interval)
+  componentDidMount () {
+    this.check()
+    Linking.addEventListener('url', this.handle)
   }
 
   componentWillUnmount () {
-    this.stopSendTokenRequest()
+    Linking.removeEventListener('url', this.handle)
   }
 
-  sendTokenRequest () {
-    if (!this.ref) return
-    console.log('Sending token request')
-    this.ref.postMessage('token')
+  _handleOpenURL (event) {
+    this.handleUrl(event.url)
+  }
+
+  async check () {
+    const url = await Linking.getInitialURL()
+    this.handleUrl(url)
+  }
+
+  handleUrl (u: ?string) {
+    console.log('Handling url', u)
+    if (!u) {
+      return
+    }
+    const parsed = url.parse(u)
+    if (parsed.protocol !== 'laundree:') {
+      return
+    }
+    if (parsed.hostname !== 'auth') {
+      return
+    }
+    const pattern = /^\/([^/]+)\/([^/]+)$/
+    const matches = parsed.path.match(pattern)
+    if (!matches) {
+      return
+    }
+    this.props.onSuccess({userId: matches[1], secret: matches[2]})
   }
 
   render () {
     return <LoadingWebView
       viewStyle={authWebView.view}
-      injectedJavaScript={this.injectedJavaScript}
-      webViewRef={ref => { this.ref = ref }}
-      onMessage={evt => this.handleMessage(evt.nativeEvent.data)}
       source={this.props.source}
-      style={authWebView.webView}
-      onLoadStart={evt => this.onLoadUrl(evt.nativeEvent.url)}/>
+      style={authWebView.webView}/>
   }
 }