+++ src/react/views/GoogleAuthWebView.js
@@ -7,6 +7,6 @@ type Props = { onAuthFailed: () => void, onSuccess: () => void }
 const GoogleAuthWebView = ({onAuthFailed, onSuccess}: Props) => <AuthWebView
   onSuccess={onSuccess}
   onAuthFailed={onAuthFailed}
-  source={{uri: `${config.laundree.host}/auth/google?mode=native-app`}}/>
+  source={{uri: `${config.laundree.host}/auth/google?mode=native-app-v2`}}/>
 
 export default GoogleAuthWebView